third_party/blink/web_tests/platform/mac-mac10.10/external/wpt/html/cross-origin-embedder-policy/cache-storage-reporting.https-expected.txt

This is a testharness.js-based test.
FAIL COEP support on document. promise_test: Unhandled rejection with value: "Report not send"
PASS COEP-Report-Only support on document.
FAIL COEP support on DedicatedWorker. promise_test: Unhandled rejection with value: "Too many reports received"
FAIL COEP-Report-Only support on DedicatedWorker. assert_equals: expected "https://web-platform.test:8444/html/cross-origin-embedder-policy/resources/universal-worker.js" but got "https://web-platform.test:8444/html/cross-origin-embedder-policy/resources/iframe.html"
FAIL COEP support on SharedWorker. promise_test: Unhandled rejection with value: "Report not send"
FAIL COEP-Report-Only support on SharedWorker. promise_test: Unhandled rejection with value: "Report not send"
FAIL COEP support on ServiceWorker. promise_test: Unhandled rejection with value: "Report not send"
PASS COEP-Report-Only support on ServiceWorker.
Harness: the test ran to completion.
